EC-519 Refactored IDeviceActionService to be split into IFileService and IAutofillManager also some cleanups were made
