+++ src/Android/MainApplication.cs
@@ -138,8 +138,9 @@ namespace Bit.Droid
             var stateMigrationService =
                 new StateMigrationService(liteDbStorage, preferencesStorage, secureStorageService);
             var clipboardService = new ClipboardService(stateService);
-            var deviceActionService = new DeviceActionService(clipboardService, stateService, messagingService,
-                broadcasterService, () => ServiceContainer.Resolve<IEventService>("eventService"));
+            var deviceActionService = new DeviceActionService(stateService, messagingService);
+            var fileService = new FileService(stateService, broadcasterService);
+            var autofillHandler = new AutofillHandler(stateService, messagingService, clipboardService, new LazyResolve<IEventService>());
             var platformUtilsService = new MobilePlatformUtilsService(deviceActionService, clipboardService,
                 messagingService, broadcasterService);
             var biometricService = new BiometricService();
@@ -158,6 +159,8 @@ namespace Bit.Droid
             ServiceContainer.Register<IStateMigrationService>("stateMigrationService", stateMigrationService);
             ServiceContainer.Register<IClipboardService>("clipboardService", clipboardService);
             ServiceContainer.Register<IDeviceActionService>("deviceActionService", deviceActionService);
+            ServiceContainer.Register<IFileService>(fileService);
+            ServiceContainer.Register<IAutofillHandler>(autofillHandler);            
             ServiceContainer.Register<IPlatformUtilsService>("platformUtilsService", platformUtilsService);
             ServiceContainer.Register<IBiometricService>("biometricService", biometricService);
             ServiceContainer.Register<ICryptoFunctionService>("cryptoFunctionService", cryptoFunctionService);
