+++ src/App/Pages/Settings/AutofillServicesPageViewModel.cs
@@ -10,6 +10,7 @@ namespace Bit.App.Pages
     public class AutofillServicesPageViewModel : BaseViewModel
     {
         private readonly IDeviceActionService _deviceActionService;
+        private readonly IAutofillHandler _autofillHandler;
         private readonly IStateService _stateService;
         private readonly MobileI18nService _i18nService;
 
@@ -22,6 +23,7 @@ namespace Bit.App.Pages
         public AutofillServicesPageViewModel()
         {
             _deviceActionService = ServiceContainer.Resolve<IDeviceActionService>("deviceActionService");
+            _autofillHandler = ServiceContainer.Resolve<IAutofillHandler>();
             _stateService = ServiceContainer.Resolve<IStateService>("stateService");
             _i18nService = ServiceContainer.Resolve<II18nService>("i18nService") as MobileI18nService;
             PageTitle = AppResources.AutofillServices;
@@ -163,7 +165,7 @@ namespace Bit.App.Pages
             }
             else
             {
-                _deviceActionService.DisableAutofillService();
+                _autofillHandler.DisableAutofillService();
             }
         }
 
@@ -193,9 +195,9 @@ namespace Bit.App.Pages
         public void UpdateEnabled()
         {
             AutofillServiceToggled =
-                _deviceActionService.HasAutofillService() && _deviceActionService.AutofillServiceEnabled();
-            AccessibilityToggled = _deviceActionService.AutofillAccessibilityServiceRunning();
-            DrawOverToggled = _deviceActionService.AutofillAccessibilityOverlayPermitted();
+                _autofillHandler.SupportsAutofillService() && _autofillHandler.AutofillServiceEnabled();
+            AccessibilityToggled = _autofillHandler.AutofillAccessibilityServiceRunning();
+            DrawOverToggled = _autofillHandler.AutofillAccessibilityOverlayPermitted();
         }
 
         private async Task UpdateInlineAutofillToggledAsync()
