+++ src/Core/Abstractions/IAutofillHandler.cs
@@ -1,0 +1,16 @@
+﻿using Bit.Core.Models.View;
+
+namespace Bit.Core.Abstractions
+{
+    public interface IAutofillHandler
+    {
+        bool AutofillServicesEnabled();
+        bool SupportsAutofillService();
+        void Autofill(CipherView cipher);
+        void CloseAutofill();
+        bool AutofillAccessibilityServiceRunning();
+        bool AutofillAccessibilityOverlayPermitted();
+        bool AutofillServiceEnabled();
+        void DisableAutofillService();
+    }
+}
