+++ src/Core/Abstractions/IFileService.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Threading.Tasks;
+
+namespace Bit.Core.Abstractions
+{
+    public interface IFileService
+    {
+        bool CanOpenFile(string fileName);
+        bool OpenFile(byte[] fileData, string id, string fileName);
+        bool SaveFile(byte[] fileData, string id, string fileName, string contentUri);
+        Task ClearCacheAsync();
+        Task SelectFileAsync();
+    }
+}
