+++ src/Core/Utilities/LazyResolve.cs
@@ -2,8 +2,13 @@
 
 namespace Bit.Core.Utilities
 {
-    public class LazyResolve<T> : Lazy<T>
+    public class LazyResolve<T> : Lazy<T> where T : class
     {
+        public LazyResolve()
+            : base(() => ServiceContainer.Resolve<T>())
+        {
+        }
+
         public LazyResolve(string containerKey)
             : base(() => ServiceContainer.Resolve<T>(containerKey))
         {
